+++ cmake/ImportSparrow.cmake
@@ -18,7 +18,7 @@ macro(import_sparrow)
       download_project(
         PROJ scine-sparrow
         GIT_REPOSITORY https://github.com/qcscine/sparrow.git
-        GIT_TAG        3.0.0
+        GIT_TAG        3.0.1
         QUIET
       )
       # Note: Options defined in the project calling this function override default