+++ CHANGELOG.rst
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+Release 5.0.0
+-------------
+
+- Update to C++17 standard
+- Add script for recursive Python doctesting
+- Add CMake import for LBFGS++
+- Improve how CMake finds MKL and Python interpreter
+- Various bugfixes and improvements
+
 Release 4.0.0
 -------------